[theme] move .*(Size|Padding) tests to size_test.go
