+++ theme/theme_test.go
@@ -46,26 +46,6 @@ func TestTheme_Light_ReturnsCorrectBackground(t *testing.T) {
 	assert.Equal(t, LightTheme().Color(ColorNameBackground, VariantLight), bg, "wrong light theme background color")
 }
 
-func Test_TextSize(t *testing.T) {
-	fyne.CurrentApp().Settings().SetTheme(DarkTheme())
-	assert.Equal(t, DarkTheme().Size(SizeNameText), TextSize(), "wrong text size")
-}
-
-func Test_Padding(t *testing.T) {
-	fyne.CurrentApp().Settings().SetTheme(DarkTheme())
-	assert.Equal(t, DarkTheme().Size(SizeNamePadding), Padding(), "wrong padding")
-}
-
-func Test_IconInlineSize(t *testing.T) {
-	fyne.CurrentApp().Settings().SetTheme(DarkTheme())
-	assert.Equal(t, DarkTheme().Size(SizeNameInlineIcon), IconInlineSize(), "wrong inline icon size")
-}
-
-func Test_ScrollBarSize(t *testing.T) {
-	fyne.CurrentApp().Settings().SetTheme(DarkTheme())
-	assert.Equal(t, DarkTheme().Size(SizeNameScrollBar), ScrollBarSize(), "wrong inline icon size")
-}
-
 func TestEmptyTheme(t *testing.T) {
 	fyne.CurrentApp().Settings().SetTheme(&emptyTheme{})
 	assert.NotNil(t, ForegroundColor())
